2010-01-27 Miguel de Dios  <miguel.dedios@artica.es> 	* include/functions_gis.php: add the prototipe of function "saveMap2" at 	the moment for to develop in next hours or days. 	* include/javascript/jquery.json.js: add json plugin for jquery. 	* godmode/gis_maps/configure_gis_map.php: rewrite all source code, change 	the form and adapt to new gis tables. 	* godmode/gis_maps/index.php: few changes for more correct source code.
git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@2312 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/godmode/gis_maps/index.php
@@ -40,18 +40,21 @@ $table->align[3] = 'center';
 $maps = get_db_all_rows_in_table ('tgis_map','map_name');
 
 $table->data = array();
-foreach ($maps as $map) {
-	$table->data[] = array('<a href="index.php?sec=gismaps&sec2=operation/gis_maps/render_view&id='.$map['id_tgis_map'].'&amp;action=edit_map">' . $map['map_name'] . '</a>',
-		print_group_icon ($map['group_id'], true),
-		'<a href="index.php?sec=gismaps&sec2=operation/gis_maps/render_view&id='.$map['id_tgis_map'].'">' . print_image ("images/eye.png", true).'</a>',
-		'<a href="index.php?sec=godgismaps&amp;sec2=godmode/gis_maps/index&amp;id_map='.$map['id_tgis_map'].'&amp;action=delete_map">' . print_image ("images/cross.png", true).'</a>'); 
+
+if ($maps !== false) {
+	foreach ($maps as $map) {
+		$table->data[] = array('<a href="index.php?sec=gismaps&sec2=operation/gis_maps/render_view&id='.$map['id_tgis_map'].'&amp;action=edit_map">' . $map['map_name'] . '</a>',
+			print_group_icon ($map['group_id'], true),
+			'<a href="index.php?sec=gismaps&sec2=operation/gis_maps/render_view&id='.$map['id_tgis_map'].'">' . print_image ("images/eye.png", true).'</a>',
+			'<a href="index.php?sec=godgismaps&amp;sec2=godmode/gis_maps/index&amp;id_map='.$map['id_tgis_map'].'&amp;action=delete_map">' . print_image ("images/cross.png", true).'</a>'); 
+	}
 }
 
 print_table($table);
 
 echo '<div class="action-buttons" style="width: '.$table->width.'">';
 echo '<form action="index.php?sec=godgismaps&amp;sec2=godmode/gis_maps/configure_gis_map" method="post">';
-print_input_hidden ('action','create_map');
+print_input_hidden ('action','new_map');
 print_submit_button (__('Create'), '', false, 'class="sub next"');
 echo '</form>';
 echo '</div>';
